bot/requirements.txt
pynacl
discord.py[voice]
pip
youtube_dl
colorlog
cffi 